Add shortcut to restore last minimied window
To keep the functionality more compatable with windows10 behaviour,
it would be preferable to use the Meta+Up action to restore the last minimized
window if no other window has taken the focus by being moved,
clicked on or typed into.
However there does not appear to be a way in the KWin API to
determine if a window has been clicked on or typed into.

This means the following situation can occur:
If you have two windows, A and B, and say A gets minimized via Meta+Down.
B will become the active window, if B is typed into the expected action from
Meta+Up is for B to be maximized, however since we do not know B had been
typed into, A will be restored instead.